backend/requirements.txt
# SOHOBI integrated semantic kernel
# Python 3.12

# AI orchestration
semantic-kernel==1.41.1
openai==2.30.0

# API server
fastapi==0.135.3
uvicorn[standard]==0.42.0
pydantic==2.11.5

# Azure
azure-identity==1.25.3
azure-search-documents==11.6.0
azure-core==1.39.0
azure-cosmos==4.15.0              # 세션 Stateless 저장소
azure-storage-blob==12.25.1      # 로그 영구 저장소

# PDF generation
reportlab==4.4.10
pypdf==6.9.2
pdfkit==1.0.0
jinja2==3.1.6

# Finance visualization
matplotlib==3.10.8
numpy==2.4.4

# PostgreSQL DB (상권 에이전트 — Azure PostgreSQL Flexible Server)
psycopg2-binary==2.9.9

# 캐싱
cachetools==5.3.3

# Rate limiting
slowapi==0.1.9

# 유틸리티
python-dotenv==1.2.2
requests==2.33.1
chardet==7.4.0.post2
aiohttp==3.13.5
httpx==0.27.2
authlib==1.3.2
python-jose[cryptography]==3.3.0
